backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
anthropic==0.34.2
google-generativeai>=0.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
pypdf2==3.0.1
python-docx==1.1.2
slowapi==0.1.9
httpx==0.28.1
aiofiles==24.1.0
chardet==5.2.0
